New changes.

+++ Repositories/MessageRepository.cs
@@ -1,8 +1,12 @@
 ﻿using ChatApp.ApplicationDbContext;
 using ChatApp.Handlers;
+using ChatApp.HubContext;
 using ChatApp.Interfaces;
 using ChatApp.Models;
+using ChatApp.Server.Hubs;
+using Microsoft.AspNetCore.SignalR;
 using Microsoft.EntityFrameworkCore;
+using System.Diagnostics;
 using System.Security.Claims;
 using System.Text.Json;
 
@@ -13,48 +17,17 @@ namespace ChatApp.Repositories
         private readonly AppDbContext _dbContext;
         private readonly IHttpContextAccessor _contextAccessor;
         private readonly IAppLogsRepository _LogsRepository;
+        private readonly IHubContext<ChatHub, IChatHub> _hubContext;
 
-        public MessageRepository(AppDbContext dbContext, IHttpContextAccessor contextAccessor, IAppLogsRepository logsRepository)
+        public MessageRepository(AppDbContext dbContext, IHttpContextAccessor contextAccessor, IAppLogsRepository logsRepository, IHubContext<ChatHub, IChatHub> messageHub)
         {
             _dbContext = dbContext;
             _contextAccessor = contextAccessor;
             _LogsRepository = logsRepository;
+            _hubContext = messageHub;
         }
 
-        //public async Task<BaseResult> GetMessages(int id)
-        //{
-        //    try
-        //    {
-        //        var userId = await _dbContext.users.Where(x => x.Id == id).FirstOrDefaultAsync();
-
-        //        if (userId != null)
-        //        {
-        //            var data = await _dbContext.messages.Where(x => x.SenderId == id || x.ReceiverId == id).ToListAsync();
-
-        //            return new BaseResult
-        //            {
-        //                Data = data,
-        //                Code = System.Net.HttpStatusCode.OK,
-        //                IsError = false
-        //            };
-        //        }
-        //        else
-        //        {
-        //            return new BaseResult
-        //            {
-        //                Data = null,
-        //                Code = System.Net.HttpStatusCode.NotFound,
-        //                IsError = true
-        //            };
-        //        }
-        //    }
-        //    catch (Exception)
-        //    {
-        //        throw;
-        //    }
-        //}
-
-        public async Task<BaseResult> Create(Message message)
+        public async Task<BaseResult> CreatePrivateMessage(Message message)
         {
             try
             {
@@ -92,6 +65,9 @@ namespace ChatApp.Repositories
                     Message = message.Content,
                     Timestamp = message.Timestamp
                 };
+                var signalRMsg = _hubContext.Clients.All.SendPrivateMessage(msg.SenderId.ToString(), msg.Message, msg.ReceiverId.ToString());
+
+                Debug.WriteLine($"SignalR message sent: {signalRMsg}");
 
                 return new BaseResult
                 {
@@ -118,7 +94,17 @@ namespace ChatApp.Repositories
             {
                 var msg = await _dbContext.messages.FirstOrDefaultAsync(c => c.Id == id);
 
-                 _dbContext.messages.Remove(msg);
+                if (msg == null)
+                {
+                    return new BaseResult
+                    {
+                        IsError = true,
+                        Code = System.Net.HttpStatusCode.NotFound,
+                        Message = "Message not found"
+                    };
+                }
+
+                _dbContext.messages.Remove(msg);
                 await _dbContext.SaveChangesAsync();
 
                 return new BaseResult
@@ -128,7 +114,7 @@ namespace ChatApp.Repositories
                     Message = "Message deleted successfully"
                 };
             }
-            catch(Exception)
+            catch (Exception)
             {
                 throw;
             }
@@ -151,6 +137,7 @@ namespace ChatApp.Repositories
                 }
 
                 oldMsg.Content = msg.Content;
+                oldMsg.Timestamp = msg.Timestamp;
                 await _dbContext.SaveChangesAsync();
 
                 var updatedeMsgDto = new
@@ -185,19 +172,35 @@ namespace ChatApp.Repositories
 
         public async Task<BaseResult> GetMessagesForUser(int userId)
         {
-            var data = await _dbContext.messages.Where(m => m.ReceiverId == userId || m.SenderId == userId).ToListAsync();
-
-            return new BaseResult
+            try
             {
-                Data = data,
-                Code = System.Net.HttpStatusCode.OK,
-                IsError = false
-            };
-        }
+                var data = await _dbContext.messages.Where(m => m.ReceiverId == userId || m.SenderId == userId).ToListAsync();
+                if (data == null || data.Count == 0)
+                {
+                    return new BaseResult
+                    {
+                        IsError = false,
+                        Code = System.Net.HttpStatusCode.NotFound,
+                        Message = "No messages found for this user"
+                    };
+                }
 
-        public async Task AddMessage(Message message)
-        {
-            await _dbContext.messages.AddAsync(message);
+                return new BaseResult
+                {
+                    Data = data,
+                    Code = System.Net.HttpStatusCode.OK,
+                    IsError = false
+                };
+            }
+            catch (Exception ex)
+            {
+                return new BaseResult
+                {
+                    IsError = true,
+                    Message = $"An error occurred while retrieving messages: {ex.Message}",
+                    Data = null
+                };
+            }
         }
 
     }
